obo_example_app/requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
aiosqlite
pydantic
